+++ CHANGELOG.md
@@ -10,6 +10,8 @@ This is the log of notable changes to EAS CLI and related packages.
 
 ### 🐛 Bug fixes
 
+- Show original GraphQL error message in case of an unexpected error. ([#1862](https://github.com/expo/eas-cli/pull/1862) by [@dsokal](https://github.com/dsokal))
+
 ### 🧹 Chores
 
 ## [3.13.2](https://github.com/expo/eas-cli/releases/tag/v3.13.2) - 2023-05-26